requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio
tensorboard
numpy
matplotlib
pillow
scikit-learn
tqdm
timm
albumentations